Extracted Declaration from factory.go
